requirements.txt
asgiref==3.4.1
boto3==1.18.58
botocore==1.21.58
Django==3.2.8
django-environ==0.7.0
django-extensions==3.1.3
django-filter==21.1
django-storages==1.12
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
importlib-metadata==4.8.1
jmespath==0.10.0
Markdown==3.3.4
Pillow==8.3.2
psycopg2==2.9.1
PyJWT==2.2.0
python-dateutil==2.8.2
pytz==2021.3
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
typing-extensions==3.10.0.2
urllib3==1.26.7
zipp==3.6.0
